% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowfunc.R
\name{mov_mean}
\alias{mov_mean}
\title{Fast implementation of moving average}
\usage{
mov_mean(
  data,
  window_size,
  type = c("ogita", "normal", "weighted", "fading"),
  eps = 0.9
)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window_size}{moving sd window size}

\item{rcpp}{a \code{logical}. Uses rcpp implementation.}
}
\value{
Returns a \code{vector} with the moving average
}
\description{
This function does not handle NA values
}
\examples{

}
