% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowfunc.R
\name{mov_std}
\alias{mov_std}
\title{Fast implementation of moving standard deviation}
\usage{
mov_std(data, window_size, rcpp = TRUE)
}
\arguments{
\item{data}{a \code{vector} or a column \code{matrix} of \code{numeric}.}

\item{window_size}{moving sd window size}

\item{rcpp}{a \code{logical}. Uses rcpp implementation.}
}
\value{
Returns a \code{vector} with the moving standard deviation
}
\description{
This function does not handle NA values
}
\examples{

}
