% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowfunc.R
\name{movmean_std}
\alias{movmean_std}
\title{Fast implementation of moving average and moving standard deviation}
\usage{
movmean_std(data, window_size, rcpp = TRUE)
}
\value{
Returns a \code{list} with \code{avg} and \code{sd} \code{vector}s
}
\description{
This function does not handle NA values
}
