% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_equal.R
\name{test_equal}
\alias{test_equal}
\title{Test similarity}
\usage{
test_equal(mpa_obj, mpb_obj, tolerance = .Machine$double.eps)
}
\arguments{
\item{mpa_obj}{First object}

\item{mpb_obj}{Second object}

\item{tolerance}{tolerance}
}
\description{
Compute if two Matrix Profiles and Profile Indexes are similar. Used only for internal purposes.
}
\keyword{internal}
