% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_equal.R
\name{test_equal}
\alias{test_equal}
\title{Test similarity}
\usage{
test_equal(mpa_obj, mpb_obj, tolerance = .Machine$double.eps)
}
\description{
Computes the best so far Matrix Profile and Profile Index for Univariate Time Series.
}
\keyword{internal}
